requirements.txt
flask~=3.0.0
python-dateutil~=2.9.0
peewee~=3.17.0